requirements.txt
python-gitlab
flask 
httplib2
python-dateutil
pyyaml
instana
